# docs/swagger.yaml
definitions:
...
  usercenter.CommonResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  usercenter.DelRequest:
    properties:
      id:
        description: '@inject_tag: validate:"required"'
        type: string
      tenantID:
        type: string
    required:
    - id
    type: object
  usercenter.EnableRequest:
    properties:
      enable:
        type: boolean
      id:
        description: '@inject_tag: validate:"required"'
        type: string
      tenantID:
        type: string
    required:
    - id
    type: object
...
  usercenter.GetTenantDetailResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/usercenter.TenantInfo'
      message:
        type: string
    type: object
...
  usercenter.QueryTenantResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/usercenter.TenantInfo'
        type: array
      message:
        type: string
      pages:
        type: integer
      records:
        type: integer
      total:
        type: integer
    type: object
...
  usercenter.TenantInfo:
    properties:
      address:
        type: string
      area:
        description: 区/县
        type: string
      areaCovered:
        type: string
      businessScope:
        type: string
      cellPhone:
        type: string
      city:
        description: 城市
        type: string
      contact:
        type: string
      enable:
        type: boolean
      id:
        type: string
      name:
        description: '@inject_tag: validate:"required"'
        type: string
      province:
        description: 省份
        type: string
      staffSize:
        type: integer
      town:
        description: 街道/镇
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
paths:
  /api/auth/tenant/export:
    get:
      consumes:
      - application/json
      description: 导出
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      - description: 租户名称
        in: query
        name: name
        type: string
      - description: IDs
        in: query
        items:
          type: string
        name: ids
        type: array
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 导出
      tags:
      - 租户管理
  /api/auth/tenant/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入
      parameters:
      - description: Bearer+空格+Token
        in: header
        name: authorization
        required: true
        type: string
      - description: 要上传的文件
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 导入
      tags:
      - 租户管理
...
  /api/core/auth/tenant/add:
    post:
      consumes:
      - application/json
      description: 新增租户
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.TenantInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 新增租户
      tags:
      - 租户管理
  /api/core/auth/tenant/all:
    get:
      consumes:
      - application/json
      description: 查询所有租户
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QueryTenantResponse'
      summary: 查询所有租户
      tags:
      - 租户管理
  /api/core/auth/tenant/delete:
    delete:
      consumes:
      - application/json
      description: 软删除租户
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.DelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 删除租户
      tags:
      - 租户管理
  /api/core/auth/tenant/detail:
    get:
      consumes:
      - application/json
      description: 查询明细
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.GetTenantDetailResponse'
      summary: 查询明细
      tags:
      - 租户管理
  /api/core/auth/tenant/enable:
    post:
      consumes:
      - application/json
      description: 禁用/启用租户
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.EnableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 禁用/启用租户
      tags:
      - 租户管理
  /api/core/auth/tenant/query:
    get:
      consumes:
      - application/json
      description: 分页查询
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      - description: 租户ID
        in: query
        name: tenantID
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QueryTenantResponse'
      summary: 分页查询
      tags:
      - 租户管理
  /api/core/auth/tenant/update:
    put:
      consumes:
      - application/json
      description: 更新租户
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.TenantInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 更新租户
      tags:
      - 租户管理
...
swagger: "2.0"